refactor Index page
